pyproject.toml
[tool.poetry]
name = "telegramtogptgateway"
version = "0.1.0"
description = ""
authors = ["Anatoly <ad.uglev@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.30"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.2.1"
greenlet = "^3.0.3"
httpx = "^0.27.0"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
redis = "^5.0.4"
passlib = "^1.7.4"
bcrypt = "4.0.1"
openai = "^1.30.1"
requests = "^2.31.0"
jose = "^1.0.0"
email-validator = "^2.1.1"
pyjwt = "^2.8.0"
aioredis = "^2.0.1"
distro-info = "1.0"
pycryptodome = "^3.20.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
